requirements.txt
ccxt>=4.1.13
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0 